initial loading/gui layout done, need to do joptionpanes for buy/sell
refactor how new tabs are opened, as removing all and readding all seems
to be able to throw an error if switching tabs as this method is called.

either, make gui freeze as tabs would only change in light of actions
through gui, and resume after update(). Or refactor entirely.

added methods to allow editing of jtable fields, however methods could
be unused if the class just makes a new table whenever theres a refresh.